requirements.txt
python-dotenv
openai
mcp
requests
beautifulsoup4
sqlalchemy
pydantic
numpy